requirements.txt
requests>=2.25.1
GitPython>=3.1.18
validators>=0.18.2
